third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-lifetime-mediakeys-with-session.html

<!DOCTYPE html>
<html>
    <head>
        <title>Test MediaKeys lifetime when adding a session</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <script>
            // MediaKeySessions remain as long as:
            //   JavaScript has a reference to it
            //   OR (MediaKeys is around
            //       AND the session has not received a close() event)
            // In the tests below, we do not close any session nor keep a
            // Javascript reference to any session, so MediaKeySessions remain
            // as long as the associated MediaKeys object is around.

            // For this test, create a MediaKeySession and verify lifetime.
            async_test(function(test)
            {
                gc();
                var initDataType;
                var initData;
                var mediaKeys;
                var startingMediaKeysCount = window.internals.mediaKeysCount();
                var startingMediaKeySessionCount = window.internals.mediaKeySessionCount();

                function numMediaKeysCreated()
                {
                    return window.internals.mediaKeysCount() - startingMediaKeysCount;
                }

                function numMediaKeySessionCreated()
                {
                    return window.internals.mediaKeySessionCount() - startingMediaKeySessionCount;
                }

                // Create a MediaKeys object with a session.
                navigator.requestMediaKeySystemAccess('org.w3.clearkey', getSimpleConfiguration()).then(function(access) {
                    initDataType = access.getConfiguration().initDataTypes[0];
                    initData = getInitData(initDataType);
                    return access.createMediaKeys();
                }).then(function(result) {
                    mediaKeys = result;

                    assert_equals(numMediaKeysCreated(), 1, 'MediaKeys.create()');
                    assert_equals(numMediaKeySessionCreated(), 0, 'After final gc()');

                    var mediaKeySession = mediaKeys.createSession();
                    return mediaKeySession.generateRequest(initDataType, initData);
                }).then(function() {
                    assert_equals(numMediaKeysCreated(), 1, 'MediaKeys.createSession()');
                    assert_equals(numMediaKeySessionCreated(), 1, 'MediaKeys.createSession()');

                    // Run gc(), should not affect MediaKeys object nor the
                    // session since we still have a reference to it.

                    // When enabling oilpan GC, the in-active
                    // ScriptPromiseResolvers will be destroyed.
                    return createGCPromise();
                }).then(function(result) {
                    assert_equals(typeof mediaKeys.createSession, 'function');

                    assert_equals(numMediaKeysCreated(), 1, 'After gc()');
                    assert_equals(numMediaKeySessionCreated(), 1, 'After gc()');

                    // Drop reference to the MediaKeys object and run gc()
                    // again. Object should be collected this time. Since
                    // MediaKeySessions remain alive as long as MediaKeys is
                    // around, it is possible that gc() checks the
                    // MediaKeySession object first, and doesn't collect it
                    // since MediaKeys hasn't been collected yet. Thus run gc()
                    // twice, to ensure that the unreferenced MediaKeySession
                    // object get collected.
                    mediaKeys = null;
                    return createGCPromise();
                }).then(function(result) {
                    return createGCPromise();
                }).then(function(result) {
                    assert_equals(numMediaKeysCreated(), 0, 'After final gc()');
                    assert_equals(numMediaKeySessionCreated(), 0, 'After final gc()');

                    test.done();
                }).catch(function(error) {
                    forceTestFailureFromPromise(test, error);
                });
            }, 'MediaKeys lifetime with session');
        </script>
    </body>
</html>
